Use Python 3.11
Python 3.8 was failing to import biosimulators_utils because it used 'list[str]' as a type hint for return values, and that syntax wasn't added until Python 3.9.  

And while I'm updating stuff, might as well move to 3.11, to see if that works.  These sorts of things will only get worse with time if we don't.  (If things fail, I'll revert to 3.10 or 3.9.)